+++ tests/hazmat/bindings/test_utils.py
@@ -13,13 +13,13 @@
 
 from __future__ import absolute_import, division, print_function
 
-from cryptography.__about__ import (
-    __author__, __copyright__, __email__, __license__, __summary__, __title__,
-    __uri__, __version__
-)
+from cryptography.hazmat.bindings import utils
 
 
-__all__ = [
-    "__title__", "__summary__", "__uri__", "__version__", "__author__",
-    "__email__", "__license__", "__copyright__",
-]
+def test_create_modulename():
+    cdef_sources = ["cdef sources go here"]
+    source = "source code"
+    name = utils._create_modulename(cdef_sources, source, "2.7")
+    assert name == "_Cryptography_cffi_bcba7f4bx4a14b588"
+    name = utils._create_modulename(cdef_sources, source, "3.2")
+    assert name == "_Cryptography_cffi_a7462526x4a14b588"