+++ tests/hazmat/primitives/test_hkdf.py
@@ -13,9 +13,9 @@
 
 from __future__ import absolute_import, division, print_function
 
-import six
-
 import pytest
+
+import six
 
 from cryptography.exceptions import (
     AlreadyFinalized, InvalidKey, UnsupportedInterface